Updated 2D files and added automated main script

+++ 2D/Helper/jacobi.h
@@ -1,12 +1,12 @@
 __host__ __device__
-float jacobi2DPoisson(const float leftX, const float rightX, const float topX, const float bottomX, const float centerRhs, const float dx, const float dy)
+double jacobi2DPoisson(const double leftX, const double rightX, const double topX, const double bottomX, const double centerRhs, const double dx, const double dy)
 {
     return (centerRhs*dx*dx*dy*dy + dy*dy*(leftX + rightX) + dx*dx*(topX+bottomX)) / (2.0*(dx*dx + dy*dy));
 }
 
 __host__ __device__
-float jacobi(const float leftMatrix, const float centerMatrix, const float rightMatrix, const float topMatrix, const float bottomMatrix,
-             const float leftX, const float centerX, const float rightX, const float topX, const float bottomX, const float centerRhs)
+double jacobi(const double leftMatrix, const double centerMatrix, const double rightMatrix, const double topMatrix, const double bottomMatrix,
+             const double leftX, const double centerX, const double rightX, const double topX, const double bottomX, const double centerRhs)
 {
     return (centerRhs - (leftMatrix * leftX + rightMatrix * rightX + topMatrix * topX + bottomMatrix * bottomX)) / centerMatrix;
 }
